Bug fixes and wired option implemented

+++ database/migrations/2021_12_21_012034_create_products_table.php
@@ -27,7 +27,7 @@ class CreateProductsTable extends Migration
             $table->integer('user_id');
             $table->integer('wireless')->default(false)->nullable();
             $table->integer('product_type')->default(1);
-            $table->float('warranty')->nullable();
+            $table->integer('warranty')->nullable();
             $table->integer('is_special')->default(0)->nullable();
             $table->integer('status')->default(true)->nullable();
             $table->text('description')->nullable();
